Only update docker-compose if a file is present

+++ src/plone_repo_helper/commands/release.py
@@ -101,7 +101,7 @@ def _step_update_repository(
     dutils.indented_print(f"- Updated {version_file} file")
     # Update docker-compose.yml
     compose_file = settings.compose_path
-    if compose_file:
+    if compose_file.exists() and compose_file.is_file():
         contents = compose_file.read_text().replace(original_version, next_version)
         compose_file.write_text(contents)
         dutils.indented_print(f"- Updated {compose_file} file")
